Server/requirements.txt
Django>=4.2,<5.0
graphene-django
psycopg2-binary
django-cors-headers
dj-database-url
python-dotenv
